chore: consistently lower case the query in at search level
At the moment, because the query is lower-cased in `def search`, there is no way to reimplement and of the `match_` methods with case sensitive search.

The query is already being lower-cased in `match_app` and `match_model`. Added lower-casing to `filter_field` and removed it from `search`.

This will amke the plugin more flexible.